Move `TgsAuthorizeAttribute` from `Controllers` namespace to `Security`
